requirements.txt
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
mutagen==1.47.0
python-dotenv==1.0.1
python-telegram-bot==21.0.1
requests==2.31.0
sniffio==1.3.1
twspace-dl==2023.7.24.1
urllib3==2.2.1
